requirements.txt
tqdm
ipykernel
matplotlib
scikit-learn
opencv-python
xgboost
pandas
mlflow
flask
dill
reportlab
gunicorn
scikit-image
seaborn
Pillow
pymongo
tensorflow==2.13.0
glrlm
imblearn
dill
from_root
supabase
evidently==0.1.58.dev0
Flask==2.0.1
Flask-OAuthlib
requests==2.26.0
tenacity
Flask-SQLAlchemy
Flask-Mail
flask-login
jwt
ctransformers
sentence-transformers
authlib
langchain
langchain-community
pypdf
openai
tiktoken
faiss-cpu
python-dotenv



-e .